hw05 java spring libraryJDBC v1.3

+++ src/main/java/ru/otus/library/dao/AuthorDao.java
@@ -8,5 +8,5 @@ public interface AuthorDao {
     void insert(Author author);
     List<Author> getAll();
     Author findByFIO(String fio);
-    Author findByID(int id);
+    Author findByID(long id);
 }
